feat: add job template plugin specifications and refactor execution context
