+++ pyproject.toml
@@ -24,6 +24,7 @@ dependencies = [
   "typer==0.12.3",
   "pyyaml==6.0.2",
   "dacite==1.9.2",
+  "pluggy==1.5.0",
 ]
 requires-python = ">= 3.8"
