+++ q8s/enums.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+
+class Target(str, Enum):
+    cpu = "cpu"
+    gpu = "gpu"
+    qpu = "qpu"
